+++ cmd/datawatch/main.go
@@ -61,7 +61,7 @@ import (
 )
 
 // Version is set at build time via -ldflags.
-var Version = "0.8.0"
+var Version = "0.8.1"
 
 var (
 	cfgPath    string